refactor: move ConditionalScript to GenericPagesComponents
